github: add CAA bot
We need external contributors to assign us copyright for the submitted
work, in case we want to update the materials license. This process is
currently performed manually, with the result stored in our internal git
server.

Use CLA Assistant lite to automate this part. The automation brings the
following components:
- a dedicated github action running each time a PR is opened or edited
- a proper CAA text for individuals and entities
- a file containing the list of individuals or entities who signed the CAA

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>