yocto: slides: "Append file example" slide: use better file names
Having a 'defconfig' file is useful as it is a practical example of a file
that might exist both in the original recipe and in the bbappend, and as
such to explain why the order in FILSPATHS matters and we need to prepend
instead of appending.

The patch file is just a more realistic example.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>