yocto: slides: introduce optional dependencies before per-recipe sysroot
The "Per-recipe sysroot" slides introduce the per-recipe sysroot but not
the main motivation it exists for, i.e. reproducible builds in presence of
automatic optional dependencies in configuration scripts.

To me this takes a significant time in a verbal explanation, which would be
more effective with a supporting slide.

Add a slide to describe optional dependences with an example found in the
oe-core recipes.

Closes: https://github.com/bootlin/training-materials/issues/253

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>