yocto: slides: use a simpler example for the :append override
FILESEXTRAPATHS is a very complex example, involving :prepend, immediate
expansion and ${THISDIR}. Let it be introduced later on when mentioning
FILESPATH and replace it with a simpler example.

The example used here is maybe not the simplest possible, but it is the
most straightforward I found in oe-core. Another option would be writing a
simple example, not taken from real code.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>