debugging: labs: add hint to make trace-cmd relay setup work with older host tool
Some trainees reported that the perf data file was empty when trying to
display it in kernel shark. The recording actually failed to be pushed
onto the host, with the following error:

Connected with 192.168.0.100:33446
cpus=2
pagesize=4096
trace-cmd: Interrupted system call
  Unsupported file version 7

This error is due to trace-cmd on the host being older enough than the
one on the target to not be able to handle the trace format version used
by the latter.

Fix this issue by forcing the trace format version to an older one.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>