debugging: labs: REALLY fix path related to kernel version
There has been a mixup between linux-6.6 mentions and linux-6.6.21
mentions for the path where to find the kernel sources and the vmlinux
file used for the training. Commit 0b73b2f2f2c3 ("debugging: labs: fix
vmlinux path") tried to fix this but in fact applied the wrong path.

Fix all paths with the correct one. While at at, use a single latex
command for the kernel version so we have only one place to update from
now on.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>