debugging: slides/labs: split ebpf labs into two parts
The slides (and the corresponding lab) about ebpf are quite long. There
is a natural split available in the lab: we can perform the BCC slides,
do the first part of ebpf lab, then study libbpf, and finish with libbpf
lab.

Apply this split to improve "brain locality". Split the slides part into
two parts. For the labs part, create two dedicated labs
Since the training schedule should not change with this update, keep the
same agenda.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>