debugging: slides/labs: re-split kernel debugging
Similarly to ebpf lab, the kernel lab is pretty big and covers a lot of
content. Re-split it as well into two parts: one part with all the
runtime checkers, and one part with kgdb/vmcore debugging. Once again,
it should not change the overall schedule of the training since the new
parts follow each other, so the agenda remains untouched.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>