embedded-linux: labs: add mention about u-boot build failure and fix
u-boot build can fail on hosts with recent python environment. The issue
is due to an API break, and it has already been identified and fixed
upstream (see [1]). Add a small mention about this in case trainees
encounter the issue as well. This commit should be removed once we
upgrade u-boot on version >= v2025.04

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>