embedded-linux: labs: add hint to fix faulty systemd unit startup
When booting the systemd software stack, we can observe that a unit
fails to start, leading the system to stay in a _degraded_ state from
systemd point of view. The issue is about systemd-network-generator
failing to parse the ip field from the kernel commandline.

Add a hint about how to fix this issue to make sure that the whole boot
is flagged as correctly done.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>