debugging: slides: add simpler syntax to use retprobes
We are current probes examples, we are not really relying on kernel
capabilities to fetch the return value from a kretprobe/uretprobe: we
are rather defining a new variable "ret" and selecting the relevant
register to get this return value. The kernel is perfectly capable of
selecting the relevant register for us, by providing directly $retval
argument. The only trick with this syntax is that one has to remember to
escape the $ prefix to make sure that it is not interpreted by the shell
in which we manipulates probes.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>