labs: sysdev: add missing dependencies for ct-ng and cfdisk
On the official Ubuntu 22.04 docker image which is quite minimal, the
gettext and python3 packages are missing. Without these packages, the
crosstool-ng build process is flooded with warnings.

Moreover, the fdisk package is also absent from the base docker image, which
means that cfdisk is not installed for the BeaglePlay labs.

Add these packages to the list of dependencies so that these warnings are
not encountered by trainees starting from minimal Ubuntu 22.04 images.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>