labs: sysdev: toolchain: use aarch64-unknown-linux-uclibc sample for BeaglePlay
The BeaglePlay lab instructions currently use the aarch64-unknown-linux-musl
sample as a starting point for building the 64-bit toolchain. However, this
sample does not exist in the version of crosstool-ng used.

Use the aarch64-unknown-linux-uclibc as a starting point instead, since the C
library will be configured to musl afterwards anyway.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>