labs/sysdev-u-boot-beagleplay: use absolute paths for ATF and DM make parameters
In the version of U-Boot used in this lab, the ATF and DM make parameters are
interpreted as relative to the build directory, and not the current directory.
This causes the build to produce broken images that contain neither TF-A nor
the DM firmware.

Fix this build command by passing ATF and DM as absolute paths instead of
relative ones.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>