yocto: labs: fix apt install list for Ubuntu 24.04
We officially support Ubuntu 24.04 for the yocto training, which does not
have the python3-distutils package anymore. This is a leftover from when
the supported distribution was Ubuntu 22.04.

The questions was raised whether python3-setuptools needs to be installed
as a replacement in order for devtool to work, but devtool worked during an
onsite training with Ubuntu 24.04 without needing to install it.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>