Makefile: build agendas twice
LaTeX, in its infinite wisdom, for misc reasons needs stuff to be
built twice. One of the reasons is the tikz current_page anchor, which
(apparently) requires LaTeX to calculate the page dimensions in the
first build, to know how to position the element in the second
build. Luckily, building agendas is fast, so that's not too much of an
issue.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>