embedded-linux: labs: Use crosstool-ng 1.27
Crosstool-ng 1.27 has been released. It fixes the issue encountered with
wget2 on Fedora distributions.

Update the toolchain lab to use this new tag and select 14.2.0 gcc.
Update accordingly the buildroot lab to match the gcc version and the
kenrel headers.

Update the slide about fetching crosstool-ng to also checkout the latest
tag.

Signed-off-by: Bastien Curutchet <bastien.curutchet@bootlin.com>