debugging: slides: remove R0 mention in perf probe example
The previous version of the slide was introducing a quite manual way of
retrieving the return value from functions probed with perf probe (ie:
instructing perf about the register to capture to get the return value).
Since then, slides have been updated to let perf perform the relevant
deduction, and so the R0 mention is not relevant anymore

Remove R0 mention to avoid any confusion or architecture-specific detail
when discussing perf probe.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>