Revert "debugging: lab-data: ebpf/libbpf_advanced: fix missing common.h in Makefile"
Common.h is indeed a dependency for both the userspace and the
kernel-space parts of the libbpf_advanced lab, however this file is not
present by default: the first instructions only refers to both c files,
common.h does not exist yet. The makefile will then generate an error
when trainees will try to check that the delivered makefile works.
Maybe some trainees won't even create a common.h and just duplicate
values/definitions between bpf and userspace part. So it is fine not to
add common.h in the makefile, and let trainees enrich the Makefile if
they choose to create a common.h.

It is fine to keep this dependency on common.h in the makefile delivered
with the labs solutions, though.

This reverts commit d3359eb6e8f8d3926c59c171956657a57f29db70.