labs/sysdev-buildroot: Make the introduction to MPC less confusing
Currently, the first two commands that trainees are supposed to test for
MPC are the volume control commands, which don't work until a fix is
applied later. This can be very confusing, especially since it's not
immediately clear that the following part fixes them.

Reorder the MPC commands so that the functioning ones are suggested first,
and explicitly state that the subsequent section fixes the volume control
issue.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>