slides: yocto-recipe-extra: replace image/ by ${D}
The first diagram presenting the task execution flow uses ${S}, ${B} and
${D} to refer to the different directories present in the WORKDIR.
For diagrams that continue showing this execution flow, keep using ${D}
instead of hardcoded "image/".

This is a bit better to create a connection between what's done in the
do_install() task (we never install to ${WORKDIR}/image, but ${D}) and
the next tasks (do_package and do_populate_sysroot).

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>