slides: fix slide selection logic based on training type
Previously, when building the slides for the "linux-kernel-beagleplay"
training, the generic "linux-kernel" content was not included. This was
due to slide selection relying on exact string matching.

To fix this, replaced the equality check with a prefix match using
`\IfBeginWith` from the `xstring` package. Now, all trainings starting
with "linux-kernel" correctly include the common kernel content.

Signed-off-by: Antoine Picard <antoine.picard@bootlin.com>