debugging: lab-data: libbpf_advanced/Makefile: fix in case KDIR not defined
If 'make' is run without setting KDIR, an empty vmlinux.h is
generated. After that, setting KDIR and running make will lead to:

  clang -Wall -Werror -Wextra -Wno-unused-parameter -target bpf -g -O2 -c trace_programs.bpf.c -o trace_programs.bpf.o
  In file included from trace_programs.bpf.c:2:
  In file included from /usr/include/bpf/bpf_helpers.h:11:
  /usr/include/bpf/bpf_helper_defs.h:86:90: error: unknown type name '__u64'
     86 | static long (* const bpf_map_update_elem)(void *map, const void *key, const void *value, __u64 flags) = (void *) 2;
        |                                                                                          ^

This is beacuse vmlinux.h is already present and thus will not be
regenerated by running 'make'.

Fix by adding vmlinux as a dependency for the vmlinux.h target.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>