slides: debugging: reorder trace-cmd examples
When discussing trace-cmd, we kind of follow a logic order, matching the
process of using trace-cmd: listing available tracers or events,
starting a recording, and finally manipulating the generated traces.

The order is a bit messed up at the end though: when mentioning traces
flush, we go back to using a specific tracer (irqsoff), and tracing
custom events. This feels odd to go back to this, so slightly reorder
things to make it feel more natural:

- move irqsoff tracing and events tracing up, before traces management
- in the "using tracers" slide, merge two similar use cases (functions
  filtering and function_graph) so we can get rid of one and make room
for the irqsoff and events tracing

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>