labs: debugging: add an explicit KDIR var check in bpf build
It is _very easy_ to forget to export the KDIR variable when running the
make command for the first time in the libbpf_advanced lab. It can lead
to a variety of errors, depending on the next commands executed by
trainees.

Prevent this error by adding an explicit check that will error the build
if KDIR is not set or set incorrectly.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>