debugging: slides: bcc: use ''' for the embedded eBPF program
In case of eBPF programs embedded in python which have some special
characters (%), there is an esacping issue. This happens among others when
doing the BCC lab.

Use ''' instead of """ for the embedded string. This will avoid a headache
when using this slide as a copy-paste starting point for writing a BCC
program.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>