treewide: update the way of including shared content
Many labs and slides include some content from the "common" directory,
using the following syntax:
  \input{../common/foo.tex}

While this syntax allows building the materials, it prevents the root
Makefile from properly replacing the __SESSION_NAME_ value that is used
as a placeholder in many .tex files. To make sure to properly fill this
placeholder in shared content, the following syntax must be use:
  \input{foo.tex}

Apply this modification on all of our materials to make sure not to keep
propagating it with more copy/pasting.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>