debugging: add beagleplay variant of labs and slides
Port the debugging training to the Beagleplay board. The most notable
change is that the Beagleplay isn't an homogeneous, dual core ARM32
board, but a heterogeneous ARM64 board based on the TI AM625 SoC,
exposing:
- a Cortex-R5
- a Cortex-M4
- four Cortex-A53

Create a new debugging-beagleplay variant of the materials. This variant
uses as much as possible the existing materials, just differentiating
the relevant points: the different build commands and configuration
files, the used peripherals, etc. The few outstanding points are the
following:
- the setup part rely on the beagleplay setup described in the kernel
  training that has been extracted to be shared between trainings
- the final lab (kexec/kdump) has too many differences to simply reuse
  the existing lab, so it has dedicated tex files.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>