labs/kernel-serial-iomem-imx93-frdm: Rework entirely this chapter
The configuration can be part of a dedicated .c file and self
contained. No need to explain everything here, let's just focus on the
parts that matter to us.

Instead, mention the use of clocks clearly (will be helpful later for
the clocks chapter that is missing).

Add a task: reading the version ID register to verify the access to the
registers before giving the snippet of code used for initializing.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>