agenda: rename trainingtype to sessiontype
There are some parts of our materials that include/exclude content
depending on the 'training' variable. This pattern is fragile as the
'training' variable does not contain the training type (eg: kernel,
yocto...) but the specific hardware used for this training (eg:
beagleplay-kernel, bbb-yocto, etc).  This leads to all those points
using the 'training' variable as a condition being forgotten each time
we add a new variant. So we would like to add a 'trainingtype' variable
exposing the actual training type ("kernel", "yocto"...) in the tex
files.

However, there is already a 'trainingtype' variable in the agendas,
representing whether the session is online or onsite. While it does not
conflict with the variable we want to introduce for slides/labs, we
would enter a danger zone by making those two concepts live in the same
variable. Rename trainingtype to 'sessiontype' to avoid, at best,
confusions, at worst, accidental collisions.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>