agenda: replace training with trainingtype
As we want to distinguish between a training type (eg: kernel, yocto)
and a training type (eg: beagleplay-kernel, bbb-yocto) in the slides and
labs instructions, we need to make sure to apply the same logic in the
agenda, where such variables already exist.

Replace training with trainingtype in agendas, to at best avoid
confusions, at worst avoid accidental collisions.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>