slides: make sure to include share content based on trainingtype
Some shared content for the debugging training is included conditionally
on the training variable. It worked well so far because we had only one
variant (for STM32MP157) AND the hardware was not part of the training
variable (so it is 'debugging' for the stm32mp157 variant of the
training).

This conditional inclusion broke when introduced the Beagleplay variant
of the training, as the corresponding training variable is
'beagleplay-debugging'. Replace training with the new trainingtype to
fix this inclusion for all variants of the debugging training.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>