common: hotfix additional content inclusion for debugging training
The debugging training has been ported to beagleplay, and unfortunately
it accidentally lost parts of its content, because the `training`
variable, used to include more details about gdb and kgdb, is not
`debugging` anymore but `debugging-beagleplay`

Make sure to include back this content for the beagleplay variant. This
is not a real fix but a hotfix, as this issue has been observed during
a training session and so need an urgent fix. The proper fix, which is
in progress, is to fix how we manage the `training` variable and to stop
embedding the training variant in it.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>