Makefile: make the 'training' variable actually be the training type in slides
In our slides, we export a 'training' variable which is in fact the
training+hardware string. This does not allow to reuse this variable to
conditionally include shared content between trainings, as we would have
to enumerate all training variants to do so.

Make sure that we export the actually training type (eg: linux-kernel,
yocto, debugging, etc). The variable is not used so much so far, except
by some shared debugging resources.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>