treewide: refactor training type and board variant management
In our slides, we export a 'training' variable which is in fact the
training+hardware string. This does not allow to reuse this variable to
conditionally include shared content between trainings, as we would have
to enumerate all training variants to do so. This also forces all
training materials to redefine for example a board variable. For
training using the same boards, the correct board value may take
different forms (stm32, stm32mp1, stm32mp157...)

Refactor all the materials parameterizing by reworking a number of
points:
- realign board names for all trainings, and make sure that all training
  labs materials are parameterized on a board (eg
  full-debugging-labs.pdf becomes full-debugging-stm32mp157-labs.pdf)
- make sure that the Makefile exports a \training variable describing
  the training type (eg: linux-kernel, embedded-linux, yocto,
  debugging...) and a \board (eg: stm32mp157, bbb, beagleplay...)
  As a consequence, remove those variables definition from materials
  sources and make those use the exported vars
- group all the materials definition for a single _training_ into the
  same .mk file.
- generate a single slide deck per training: this slide deck now
  contains one shopping list per compatible board for this training

Aside from some materials generated files name, the paths involved in
some labs (those that did not have a board name attached) and the new
shopping lists, the overall content of the training materials is
unchanged.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>