Makefile: define BOARD_SUFFIXES globally, not only for slides
The BOARD_SUFFIXES variable is currently only defined if we are building
slides, and is used to handle the generic training type and/or the
specific variant. The same management is desired when building labs, but
the BOARD_SUFFIXES needs to be redefined here as well. Rather than
redefining it, properly share it and expose it for any make run

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>