slides: generate a single slide set per training type
The materials build system currently allows to generate one slide
variant per supported hardware for the labs. For example, we then end up
with:
- full-linux-kernel-bbb-slides.pdf
- full-linux-kernel-beagleplay-slides.pdf
- full-linux-kernel-imx93-frdm-slides.pdf
- etc

The only variation in those slides are in the shopping list about needed
hardware to run the labs. It does not really make sense to maintain
dedicated variants of the slides, as the concepts are the same for all
variants.

Unify the slides generation to have only one slide set per training.
Redefine standard shopping list per board, and include all the supported
shopping list boards in the training slides. Keep variants of the
shopping lists, depending on the training needs. For example, we now
have:
- shopping-list-stm32mp157
- shopping-list-stm32mp157-nunchuk (includes nunchuk ref and wires)
- shopping-list-stm32mp157-nunchuk-audio (includes audio headset/adapter)

As a side effect, all the labs for a specific training are now all
gathered in the training-specific makefile, instead of having one
makefile per lab variant.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>