slides/yocto-licensing: switch to SPDX3.0
Thanks to Kamel[1], we can now use SPDX3.0 on Scarthgap. It's overall
what most people should use nowadays as it's more featureful and simpler
to use.

Indicate that SPDX2.2 is still the default and give instructions to
switch to SPDX3.0. Mention that a single SPDX file is generated for the
image. Refresh the example to show an entry for attr, and a patch
applied to it.

[1]: https://lore.kernel.org/r/20251107131502.3857600-1-kamel.bouhara@bootlin.com

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>