common: strace: add the --tips argument to learn new commands
strace comes with a ton of features to trace app interactions with the
system, so it is easy to just miss existing features. Reading carefully
the manpage is a potential solution for the hightly motivated
developers, but strace also provides a nice --tips option, that hints
you about some powerful ways of using strace.

Mention the `strace --tips` command on strace slide.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>