Makefile: properly detect missing lab files
When computing LABS_TEX we do a $(wildcard ...) to eliminate
non-existing files. Which means that if you put a non existing lab
chapter in the <training>_<board>_LABS variable in the mk/ file, it's
simply ignored.

And this $(wildcard) call makes the check on LABS_TEX a few lines
later totally useless.

Therefore, drop this bogus $(wildcard) call. If LABS_TEX contains
non-existing files, it will be detected a few lines later.

And it turns out that indeed, a number of lab chapters were incorrect.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>