lab-data/embedded-linux-stm32mp2: share appdev with beagleplay
The "appdev" lab-data of stm32mp2 is currently shared with the one
from stm32mp1, but this is incorrect. Indeed, the stm32mp1 appdev
folder contains a .vscode folder with references to an ARM32
toolchain, but stm32mp2 is ARM64. Therefore, re-using the appdev
directory from beagleplay, which is also ARM64, will lead to a
properly working lab.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>