lab-data/embedded-linux-stm32mp1: use more generic cpuinfo CGI script
The current cpuinfo CGI script was working OK on ARM32 bit platforms
as it was grepping for the "model name" field. But this field doesn't
exist on ARM64.

Our BeaglePlay port had an entire copy of the tinysystem lab data just
for that reason, implementing a more generic solution consisting in
dumping the entire contents of /proc/cpuinfo. Since this works also
for ARM32, import this solution in embedded-linux-stm32mp1 as the
"generic" one, so that it can be used by all other platforms as well.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>